relocate to util
