+++ src/test/java/com/cjcrafter/foliascheduler/util/VersionTest.java
@@ -1,4 +1,4 @@
-package com.cjcrafter.foliascheduler;
+package com.cjcrafter.foliascheduler.util;
 
 import org.junit.jupiter.api.Test;
